refactor: 레포지토리의 위치 구조 변경
